Allow the build to work in cases where the current repo has no tags
When forking the project on GitHub there is an option to only fork the
main repo branch. This means that the forked repo will not have any tags
in it which currently causes the build system to fail.

This change adds a fallback in the case that there are no tags to
building a version named "-dev-dirty".